+++ SwiftExampleTests/SwiftExampleTests.swift
@@ -7,29 +7,21 @@
 //
 
 import XCTest
+@testable import SwiftExample
 
 class SwiftExampleTests: XCTestCase {
-    
-    override func setUp() {
-        super.setUp()
-        // Put setup code here. This method is called before the invocation of each test method in the class.
+
+    func testSafeArtworkURLAcceptsHTTPSMZStaticHosts() {
+        let controller = SearchResultsViewController()
+        XCTAssertNotNil(controller.safeArtworkURLFromString("https://mzstatic.com/artwork.png"))
+        XCTAssertNotNil(controller.safeArtworkURLFromString("https://is1-ssl.mzstatic.com/artwork.png"))
     }
-    
-    override func tearDown() {
-        // Put teardown code here. This method is called after the invocation of each test method in the class.
-        super.tearDown()
+
+    func testSafeArtworkURLRejectsUntrustedSchemesAndHosts() {
+        let controller = SearchResultsViewController()
+        XCTAssertNil(controller.safeArtworkURLFromString("http://is1-ssl.mzstatic.com/artwork.png"))
+        XCTAssertNil(controller.safeArtworkURLFromString("https://example.com/artwork.png"))
+        XCTAssertNil(controller.safeArtworkURLFromString("not a url"))
     }
-    
-    func testExample() {
-        // This is an example of a functional test case.
-        XCTAssert(true, "Pass")
-    }
-    
-    func testPerformanceExample() {
-        // This is an example of a performance test case.
-        self.measureBlock() {
-            // Put the code you want to measure the time of here.
-        }
-    }
-    
+
 }